partial implementation of request form & hashing
