+++ Explorer/Assets/DCL/Profiles/DCL.Profiles.asmdef
@@ -14,7 +14,8 @@
         "GUID:101b8b6ebaf64668909b49c4b7a1420d",
         "GUID:5eabe9a3d4dd19d42a16208ea5411062",
         "GUID:56e8195b069a4dca9c4c4f313c65f526",
-        "GUID:b46779583a009f04ba9f5f31d0e7e6ac"
+        "GUID:b46779583a009f04ba9f5f31d0e7e6ac",
+        "GUID:5ab29fa8ae5769b49ab29e390caca7a4"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
