+++ Explorer/Assets/DCL/Profiles/IProfileRepository.cs
@@ -5,6 +5,7 @@ namespace DCL.Profiles
 {
     public interface IProfileRepository
     {
+        UniTask SetAsync(Profile profile, CancellationToken ct);
         UniTask<Profile?> GetAsync(string id, int version, CancellationToken ct);
     }
 }
