+++ Explorer/Assets/Plugins/IPFS.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13b04f628b78e4cfe867e8c7bc9133a7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
