+++ Explorer/Assets/Plugins/IPFS/Common.Logging.Core.XML.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 96ed70685ec6f4f7285b6d37020f4767
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
