+++ Explorer/Assets/Plugins/IPFS/Ipfs.Core.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f1be494a8012d494b9eb5e56a3484142
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
